combined GeneralServiceIT and GeneralServiceTest
