+++ src/test/java/nl/cerios/testutil/DatabaseTest.java
@@ -17,12 +17,15 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.mockito.Spy;
 
-public class DatabaseIT {
+public class DatabaseTest {
 
+	private static final String OPTION_CREATE = "create";
+	private static final String OPTION_DROP = "drop";
+	private static final String OPTION_SHUTDOWN = "shutdown";
 	private static final Properties PROPS;
 
 	static {
-		final URL propertyDataURL = DatabaseIT.class.getResource("/integration-test.properties");
+		final URL propertyDataURL = DatabaseTest.class.getResource("/integration-test.properties");
 		try (final InputStream propertyDataStream = propertyDataURL.openStream()) {
 			PROPS = new Properties();
 			PROPS.load(propertyDataStream);
@@ -34,9 +37,14 @@ public class DatabaseIT {
 	@Spy
 	private DataSource dataSource = getDataSource();
 
-	private static DataSource getDataSource() {
+	private static DataSource getDataSource(final String... dbOptions) {
+		String dbName = "memory:" + PROPS.getProperty("db.name");
+		for (final String dbOption : dbOptions) {
+			dbName += ";" + dbOption + "=true";
+		}
+
 		final EmbeddedDataSource40 derbyDS = new EmbeddedDataSource40();
-		derbyDS.setDatabaseName("memory:" + PROPS.getProperty("db.name"));
+		derbyDS.setDatabaseName(dbName);
 		derbyDS.setUser(PROPS.getProperty("db.user"));
 		derbyDS.setPassword(PROPS.getProperty("db.password"));
 		return derbyDS;
@@ -44,21 +52,17 @@ public class DatabaseIT {
 
 	@BeforeClass
 	public static void initDatabase() throws IOException, SQLException {
-		final DataSource ds = getDataSource();
 		// tell Derby to create the schema
-		((EmbeddedDataSource40) ds).setCreateDatabase("create");
-
-		// run SQL scripts
+		final DataSource ds = getDataSource(OPTION_CREATE);
 		try (final Connection connection = ds.getConnection();
 				final Statement sqlStatement = connection.createStatement()) {
 
-			final String[] scriptResourceNames = new String[]{
-				PROPS.getProperty("db.script.create"),
-				PROPS.getProperty("db.script.testData")
-			};
-			
+			// run SQL scripts
+			final String[] scriptResourceNames = new String[] { PROPS.getProperty("db.script.create"),
+					PROPS.getProperty("db.script.testData") };
+
 			for (String scriptResourceName : scriptResourceNames) {
-				try (InputStream scriptStream = DatabaseIT.class.getResourceAsStream(scriptResourceName)) {
+				try (InputStream scriptStream = DatabaseTest.class.getResourceAsStream(scriptResourceName)) {
 					runSqlScript(sqlStatement, scriptStream);
 				}
 			}
@@ -81,14 +85,20 @@ public class DatabaseIT {
 	@AfterClass
 	public static void shutdownDatabase() throws SQLException {
 		try {
-			final DataSource ds = getDataSource();
-			((EmbeddedDataSource40) ds).setShutdownDatabase("shutdown");
-			ds.getConnection();
+			// Remove the database schema from memory...
+			getDataSource(OPTION_DROP).getConnection();
+			// ...and stop the database instance.
+			getDataSource(OPTION_SHUTDOWN).getConnection();
 		} catch (SQLNonTransientConnectionException e) {
 			// Wow, this is a dirty tricky piece of code! See
 			// http://stackoverflow.com/questions/2307788/how-to-shutdown-derby-in-memory-database-properly
 			final String dbState = e.getSQLState();
 			if (!dbState.equals("XJ015") && !dbState.equals("08006")) {
+				throw e;
+			}
+		} catch (SQLException e) {
+			// Database was already shut down.
+			if (!e.getSQLState().equals("XJ004")) {
 				throw e;
 			}
 		}
